ImGuiOverlays: Fix analog input display

+++ pcsx2/SIO/Pad/PadNotConnected.cpp
@@ -76,7 +76,7 @@ void PadNotConnected::SetAnalogInvertR(bool x, bool y)
 
 }
 
-u8 PadNotConnected::GetEffectiveInput(u32 index) const
+float PadNotConnected::GetEffectiveInput(u32 index) const
 {
 	return 0;
 }
